% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqbuild.R
\name{eq_mixed}
\alias{eq_mixed}
\title{build a formula for multilevel/mixed model based on lists of variables
   for fixed and random effects}
\usage{
eq_mixed(dv, fev, rev, rec, correlated = TRUE, addcons = FALSE)
}
\arguments{
\item{dv}{string containing the dependent variable name}

\item{fev}{vector of strings containing variable names for fixed effects}

\item{rev}{list of strings containing variable names for random effects.
Each list entry sould be a string vector containing variables
combined using the standard operators +,:,* and ^}

\item{rec}{list of strings containing factor variable names for random effects.
Each list entry sould be a single string containing factor variables
combined using the standard operators +,:,*,^ and /}

\item{correlated}{list of logicals - if TRUE the effects of re variables
are modeled as correlated.  FALSE is equivalent to the operator ||}

\item{addcons}{logical, if TRUE add a constant term to the fixed effects.
Overrides 0 and -1 in fev}
}
\value{
string
}
\description{
\code{eq_mixed} build a formula for multilevel/mixed model
}
